Fixed spacing issues and broke out the tracklistFormListener from main.
