+++ client/listeners/index.js
@@ -2,10 +2,10 @@ const attachUrlFormListener = require('./urlSubmitForm')
 const attachAddTrackButtonListener = require('./addTrackButton')
 
 function attachListeners() {
-	return {
-		urlSubmitForm: attachUrlFormListener(),
-		addTrackButton: attachAddTrackButtonListener(),
-	}
+  return {
+    urlSubmitForm: attachUrlFormListener(),
+    addTrackButton: attachAddTrackButtonListener()
+  }
 }
 
 module.exports = attachListeners
